服务器为了兼容64位做了一些调整，添加了ReadMe文件

+++ src/server.cpp
@@ -958,7 +958,7 @@ void *server_receive(void *arg)
  */
 void *server_send(void *arg)
 {
-	int sockfd = (int )arg;
+	int sockfd = (long)arg;
 	struct s_key key;
 	int status;
 	int ret;
